Added base structure
This includes the basics of the physics/timestep system, to understand what it is based on read this (https://gafferongames.com/post/fix_your_timestep/) article. I would like to see how its working, but that would be difficult without graphics. Sorry it took so long, been busier than expected this summer.